Remove duplicate Exporter and move tests to correct module
The Exporter class existed in two locations:
- feature/main/src/.../main/data/Exporter.kt (unused)
- feature/memos/data/src/.../memos/data/export/Exporter.kt (used)

Changes:
- Deleted the unused duplicate in feature/main/data
- Moved ExporterTest to feature/memos/data module where Exporter lives
